% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{run_neighbornet_networkx}
\alias{run_neighbornet_networkx}
\title{Run NeighborNet and return a networkx-like list.}
\usage{
run_neighbornet_networkx(
  x,
  flip_y = TRUE,
  labels = NULL,
  max_iterations = 5000,
  ordering_method = NULL,
  inference_method = NULL
)
}
\description{
Run NeighborNet and return a networkx-like list.
}
