% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{set_anon_nnt_threads}
\alias{set_anon_nnt_threads}
\title{Set the global thread count for anon_nnt (call once per session).}
\usage{
set_anon_nnt_threads(threads)
}
\description{
Set the global thread count for anon_nnt (call once per session).
}
